Split run_real and run_virtual methods into separate files
