+++ yash-env/src/system/concurrency.rs
@@ -525,7 +525,8 @@ impl SignalList {
 }
 
 mod delegates;
-mod run;
+mod run_real;
+mod run_virtual;
 mod rw_all;
 mod signal;
